refactoring the nabu code a bit (a lot) to help with group maintenance
