+++ internal/services/tika/tikaSingle.go
