+++ pkg/cli/root.go
@@ -1,15 +1,16 @@
 package cli
 
 import (
-	"github.com/gleanerio/nabu/internal/objects"
-	"github.com/gleanerio/nabu/pkg/config"
-	"github.com/minio/minio-go/v7"
-	"github.com/spf13/cobra"
 	"log"
 	"mime"
 	"os"
 	"path"
 	"path/filepath"
+
+	"github.com/gleanerio/nabu/internal/objects"
+	"github.com/gleanerio/nabu/pkg/config"
+	"github.com/minio/minio-go/v7"
+	"github.com/spf13/cobra"
 
 	"github.com/spf13/viper"
 )
@@ -35,7 +36,6 @@ var rootCmd = &cobra.Command{
 // Execute adds all child commands to the root command and sets flags appropriately.
 // This is called by main.main(). It only needs to happen once to the rootCmd.
 func Execute() {
-
 	cobra.CheckErr(rootCmd.Execute())
 }
