+++ pkg/textai.go
@@ -3,7 +3,7 @@ package pkg
 import (
 	"fmt"
 	"github.com/gleanerio/nabu/internal/objects"
-	"github.com/gleanerio/nabu/internal/semsearch"
+	"github.com/gleanerio/nabu/internal/services/txtai"
 	"github.com/minio/minio-go/v7"
 	"github.com/spf13/viper"
 	"log"
@@ -19,7 +19,7 @@ func NabuTxtai(v1 *viper.Viper) error {
 
 func Txtai(v1 *viper.Viper, mc *minio.Client) error {
 	fmt.Println("Index descriptions to txtai")
-	err := semsearch.ObjectAssembly(v1, mc)
+	err := txtai.ObjectAssembly(v1, mc)
 	if err != nil {
 		log.Println(err)
 	}
